requirements.txt
llama-index
pinecone-client
chainlit


